requirements.txt
pydantic>=2.0.0
ipython>=8.0.0
langchain-openai>=0.0.5
langchain-core>=0.1.0
langgraph>=0.0.10
langchain-community>=0.0.10
python-dotenv>=1.0.0
langchain-groq>=0.0.1
openai>=1.0.0
tavily-python>=0.1.0
numpy>=1.24.0
pandas>=2.0.0
Flask>=2.0.0
markdown2>=2.4.0
gunicorn>=20.0.0

